requirements.txt
pyyaml
pymodbus
twisted
requests